@@ -17,18 +17,18 @@
     "@3kles/kles-material-dialog": "^18.0.0",
     "@3kles/kles-material-dynamicforms": "^18.0.1",
     "@3kles/kles-material-datepicker": "18.0.0",
-    "@angular/animations": "^17.2.0",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.2.0",
-    "@angular/common": "^17.2.0",
-    "@angular/compiler": "^17.2.0",
-    "@angular/core": "^17.2.0",
-    "@angular/forms": "^17.2.0",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.2.0",
     "@angular/material-moment-adapter": "^17.2.0",
     "@angular/material-luxon-adapter": "^17.2.0",
-    "@angular/platform-browser": "^17.2.0",
-    "@angular/platform-browser-dynamic": "^17.2.0",
-    "@angular/router": "^17.2.0",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@infor-up/m3-odin": "^4.0.0",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
@@ -45,11 +45,11 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.0",
+    "@angular-devkit/build-angular": "^18.2.12",
     "@angular-devkit/build-ng-packagr": "^0.901.15",
-    "@angular/cli": "^17.2.0",
-    "@angular/compiler-cli": "^17.2.0",
-    "@angular/language-service": "^17.2.0",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/language-service": "^18.2.13",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.168",
@@ -62,10 +62,10 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^17.0.0",
+    "ng-packagr": "^18.2.1",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
