@@ -8,7 +8,7 @@ import { KlesMaterialTableModule } from 'kles-material-table';
 import { MaterialModule } from './modules/material.module';
 import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FakeApiService } from './services/fakemi.service';
 import { KlesMaterialDynamicformsModule } from '@3kles/kles-material-dynamicforms';
 import { KlesMaterialDialogModule } from '@3kles/kles-material-dialog';
@@ -19,39 +19,29 @@ import { TableModule } from './modules/table/table.module';
 import { TreeTableModule } from './modules/treetable/treetable.module';
 import { SharedModule } from './modules/shared/shared.module';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    AutocompleteComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    MaterialModule,
-    KlesMaterialDialogModule,
-    KlesMaterialDynamicformsModule,
-    KlesMaterialTableModule,
-    BrowserAnimationsModule,
-    HttpClientModule,
-    SharedModule,
-    TranslateModule.forRoot(
-      {
-        loader: {
-          provide: TranslateLoader,
-          useFactory: (HttpLoaderFactory),
-          deps: [HttpClient]
-        }
-      }
-    ),
-    TableModule,
-    LazyTableModule,
-    TreeTableModule,
-    LazyTreeTableModule,
-
-  ],
-  providers: [TranslateService, FakeApiService],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        AutocompleteComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        MaterialModule,
+        KlesMaterialDialogModule,
+        KlesMaterialDynamicformsModule,
+        KlesMaterialTableModule,
+        BrowserAnimationsModule,
+        SharedModule,
+        TranslateModule.forRoot({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: (HttpLoaderFactory),
+                deps: [HttpClient]
+            }
+        }),
+        TableModule,
+        LazyTableModule,
+        TreeTableModule,
+        LazyTreeTableModule], providers: [TranslateService, FakeApiService, provideHttpClient(withInterceptorsFromDi())] })
 
 export class AppModule { }
 
